requirements.txt
langchain>=0.0.267
langchain-community>=0.0.10
langchain-cohere>=0.0.1
langchain-core>=0.1.0
langchain-text-splitters>=0.0.1
langchain-chroma>=0.0.1
cohere>=4.37
chromadb>=0.4.18
pypdf>=3.17.0
python-dotenv>=1.0.0
pydantic>=2.5.0
jupyter>=1.0.0
matplotlib>=3.7.0
pandas>=2.0.0
numpy>=1.24.0